projects/pip-webui2-nav/src/lib/nav-icon/nav-icon.component.spec.ts
import * as assert from 'assert';
import { TestBed, inject, ComponentFixture } from '@angular/core/testing';
import { By } from '@angular/platform-browser';
import { DebugElement } from '@angular/core';

import { PipNavIconModule } from './nav-icon.module';
import { PipNavPartService } from '../navpart/shared/navpart.service';

import { PipNavIconComponent } from './nav-icon.component';
import { NavIconConfig } from './shared/nav-icon.model';

describe('a pip-nav-icon component', () => {
	let component: PipNavIconComponent;
	let service: PipNavPartService;
	let fixture: ComponentFixture<PipNavIconComponent>;
	let de: DebugElement;
	let el: HTMLElement;
	let navPartName = 'icon';

	// register all needed dependencies
	beforeEach(() => {
		TestBed.configureTestingModule({
			imports: [
				PipNavIconModule
			],
			providers: [
				PipNavIconComponent,
				PipNavPartService
			]
		});
	});

	// instantiation through framework injection
	beforeEach(inject([PipNavPartService], (navPartSErvice: PipNavPartService) => {
		fixture = TestBed.createComponent(PipNavIconComponent);
		service = navPartSErvice;
		component = fixture.componentInstance;

		de = fixture.debugElement;
		el = de.nativeElement;
	}));

	it('should add and show icon', () => {
		let props = { 
			icon: 'menu'
		};

		component.pipNavPartName = navPartName;
		service.updatePartByName(navPartName, true, props);
		expect(component.icon).toEqual(props);
		fixture.detectChanges();
		let iconItems = el.getElementsByTagName('mat-icon');

		expect(iconItems.length).toEqual(1);
		expect(iconItems[0].innerHTML).toEqual(props.icon);
	});
});
